Fix rename command
Renaming a project to a name that already exists will make Autumn prompt the user if they would like to merge the projects. This fixes a bug where the pre-existing project would have just been 'deleted'  when the rename command was called. This merging works for both subproject renames and project renames.